RN-972: Refactored SurveyEditor into own module - Moved utilty functions into their own files
