+++ packages/central-server/src/apiV2/surveys/EditSurvey.js
@@ -4,7 +4,7 @@
  */
 
 import { EditHandler } from '../EditHandler';
-import { SurveyEditor } from './SurveyEditor';
+import { SurveyEditor } from './surveyEditor';
 import { convertNamesToIds } from './convertNamesToIds';
 
 /**
