+++ packages/central-server/src/apiV2/surveys/surveyEditor/index.js
@@ -1,0 +1,6 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ */
+
+export { SurveyEditor } from './SurveyEditor';
